Serialization / Deserialization with Serpent
Define a __serialize function in the class metatable so that Serpent
will call it instead of serializing the instance as is.
That function adds a __deserialize function that sets the __class (a
copy of __index) as metatable and removes itself from the instance to
avoid multiple calls.
The unfortunate side effect is the copy of the metatable into the table
itself since Serpent doesn't serialize metatables.